+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add `BakeNoteIFrames` direction (minor)
 * Selector optimization in precalculus `Strategy` & elsewhere (patch)
 * Tweaks inline list seperators to only target labeled items (major?)
+* Add definition of handbook page selector needed for BakeToc, create a link class for this page (minor)
 
 ## [9.0.0] - 2021-07-12